compilation du code sur machine 64-bits:  attention dans le fichier inputfile/composition.txt le timerticker < à 40000000 n'est pas pris en compte dans cette version d'algo
execution srcipt
modification des fichiers:
+format_trace.cpp ( rajout de condition d'arret si l'appel system ne se fait pas correctement)
+generation.cpp (les lignes seekg(-line.size(),ios::cur) sont remplacées par seekg((int)(-line.size(),ios::cur) ) on a fait uncast :p
+utils.cpp modification du lancement de script d'exploration.

fait le 29 juin 2012 par J.L.S

+++ regularisation.cpp
@@ -702,7 +702,7 @@ int getSpeedTickTimer(string input)
 	if (pos!=-1)
 	{/// on  va se repositionner au debut de la ligne
 // 	cout<<line<<endl;
-	    inputfile.seekg(-line.size()-1,ios::cur);
+	    inputfile.seekg((int)(-line.size()-1),ios::cur);
 	    inputfile>>mot; 
 // 	    cout<<mot<<endl;
 	    while(mot.compare("\n")!=0 && i++<9)
@@ -729,6 +729,7 @@ int getSpeedTickTimer(string input)
 	if(nanoseconde>=600000000 && nanoseconde<1000000000) return IS_HIGH;
 	if(nanoseconde>=400000000 && nanoseconde<600000000) return IS_MEDIUM;
 	if(nanoseconde<400000000) return IS_LOW;
+	
       }
 
 }
